AI-Based-Tutor/requirements.txt
# pip install -r requirements.txt

flask
flask-cors
requests
python-dotenv